# PERF: Optimise query for excluding topics in certain categories in TopicsFilter